Fix spacemaconfig_path and update to work on Win32
I think '__file__' only worked because I was running from the directory
the script was in. It's actually a variable and not a string.

Fix existence detection since it could be a directory and not a file:
use exists() instead.

Python doesn't have os.symlink on Windows. There are a bunch of
solutions including ones that require changing permissions of
python.exe. Instead, I'm using one that shells out to mklink and works
when run from an admin command prompt (the standard mklink requirement).
Source: http://stackoverflow.com/a/28382515/79125

Tested in Win8.