Add evil-textobj-anyblock
While I don't use this in vim, it seems like a useful fallback to have
in spacemacs where I don't have all of my textobjs.

I thought 'a' would be good for 'any', but that will replace 'in angle brackets'. Instead, I'm using '.' which is like the any character regexp atom.